Revert "To avoid hiding lost columns at the bottom of the dataset"
